+++ lib/mariposa.dart
@@ -1,3 +1,5 @@
+export 'package:html_builder/html_builder.dart';
 export 'src/abstract_element.dart';
 export 'src/render_context.dart';
+export 'src/stateful_widget.dart';
 export 'src/widgets.dart';