+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0-alpha
+* Added `StatefulWidget` and `State`.
+* `package:mariposa/mariposa.dart` now exports `package:html_builder`.
+* Use `package:angel_container` in `RenderContext`.
+
 # 1.0.1
 * Add `RenderContext` and `ContextAwareWidget`.